+++ server/src/main/java/org/owasp/securityshepherd/controller/LoginController.java
@@ -42,7 +42,7 @@ public class LoginController {
 
   @PostMapping(path = "/register")
   @ResponseStatus(HttpStatus.CREATED)
-  public Mono<Integer> register(@Valid @RequestBody final PasswordRegistrationDto registerDto) {
+  public Mono<Long> register(@Valid @RequestBody final PasswordRegistrationDto registerDto) {
 
     return userService.createPasswordUser(registerDto.getDisplayName(), registerDto.getUserName(),
         passwordEncoder.encode(registerDto.getPassword()));
