+++ server/src/main/java/org/owasp/securityshepherd/repository/CorrectionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 
 @Repository
-public interface CorrectionRepository extends ReactiveCrudRepository<Correction, Integer> {
+public interface CorrectionRepository extends ReactiveCrudRepository<Correction, Long> {
   @Query("SELECT * from correction WHERE user_id = :user_id")
   public Flux<Correction> findAllByUserId(@Param("user_id") final int userId);
 }
