+++ server/src/main/java/org/owasp/securityshepherd/repository/ModulePointRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 
 @Repository
-public interface ModulePointRepository extends ReactiveCrudRepository<ModulePoint, Integer> {
+public interface ModulePointRepository extends ReactiveCrudRepository<ModulePoint, Long> {
   @Query("SELECT * from module_point WHERE module_id = :module")
   public Flux<ModulePoint> findAllByModuleId(@Param("module") final int moduleId);
 }
