+++ server/src/main/java/org/owasp/securityshepherd/repository/PasswordAuthRepository.java
@@ -11,15 +11,14 @@ import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Mono;
 
 @Repository
-public interface PasswordAuthRepository extends ReactiveCrudRepository<PasswordAuth, Integer> {
-
+public interface PasswordAuthRepository extends ReactiveCrudRepository<PasswordAuth, Long> {
   @Query("SELECT * from password_auth WHERE login_name = :login_name")
   public Mono<PasswordAuth> findByLoginName(@Param("login_name") final String loginName);
 
   @Query("SELECT * from password_auth WHERE user_id = :user_id")
-  public Mono<PasswordAuth> findByUserId(@Param("user_id") final int userId);
+  public Mono<PasswordAuth> findByUserId(@Param("user_id") final long userId);
 
   @Modifying
   @Query("delete from password_auth WHERE user_id = :user_id")
-  public Mono<UserAuth> deleteByUserId(@Param("user_id") final int userId);
+  public Mono<UserAuth> deleteByUserId(@Param("user_id") final long userId);
 }
