+++ server/src/main/java/org/owasp/securityshepherd/service/SubmissionService.java
@@ -27,7 +27,7 @@ public final class SubmissionService {
 
   private final Clock clock;
 
-  public Mono<Submission> submit(final int userId, final int moduleId, final String flag) {
+  public Mono<Submission> submit(final Long userId, final Long moduleId, final String flag) {
     if (userId <= 0) {
       return Mono.error(new InvalidUserIdException());
     }
@@ -47,7 +47,7 @@ public final class SubmissionService {
         .map(SubmissionBuilder::build).flatMap(submissionRepository::save);
   }
 
-  public Mono<Correction> submitCorrection(final int userId, final int amount,
+  public Mono<Correction> submitCorrection(final Long userId, final int amount,
       final String description) {
     if (userId <= 0) {
       return Mono.error(new InvalidUserIdException());
@@ -63,7 +63,7 @@ public final class SubmissionService {
     return correctionRepository.save(correctionBuilder.build());
   }
 
-  public Flux<Submission> findAllByModuleId(final int moduleId) {
+  public Flux<Submission> findAllByModuleId(final Long moduleId) {
     if (moduleId <= 0) {
       return Flux.error(new InvalidModuleIdException());
     }
